requirements.txt
python-dotenv
plotly
flask
pandas
kaleido